changed header size for the program to 16 bit

+++ src/generator.cpp
@@ -10,7 +10,7 @@ Generator::Generator(std::string fileName, std::string content) {
 }
 
 void Generator::generate(void) {
-    std::vector<i16> program;
+    std::vector<u16> program;
 
     while (current.type != EndOfFile) {
         program.push_back(generateNext());
@@ -18,13 +18,13 @@ void Generator::generate(void) {
     }
 
     std::ofstream file("out", std::ios::out | std::ios::binary);
-    size_t size = program.size();
+    u16 size = program.size();
     file.write((char*) &size, sizeof(size));
-    file.write((char*) &program[0], program.size() * sizeof(i16));
+    file.write((char*) &program[0], program.size() * sizeof(u16));
     file.close();
 }
 
-i16 Generator::generateNext(void) {
+u16 Generator::generateNext(void) {
     switch (current.type) {
         case Identifier:
             return lookupTable[current.value];
